right fitting instruction size

+++ core/vm/loader.h
@@ -72,8 +72,8 @@ class Loader {
     return value;
   }
   
-  inline int ReadShort() {
-    uint16_t value = *((uint16_t*)buffer);
+  inline int ReadByte() {
+    uint8_t value = *((uint8_t*)buffer);
     buffer += sizeof(value);
     return value;
   }
